+++ src/__tests__/ProgressControl.spec.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { shallow } from 'enzyme';
+import { shallow, mount } from 'enzyme';
 import ProgressControl from '../components/control-bar/ProgressControl';
 
 describe('ProgressControl', () => {
@@ -12,5 +12,10 @@ describe('ProgressControl', () => {
   it('should render with "video-react-progress-control" class', () => {
     const wrapper = shallow(<ProgressControl actions={{}} player={{}} />);
     expect(wrapper.hasClass('video-react-progress-control')).toBe(true);
+  });
+
+  it('should bind ref "seekBar"', () => {
+    const wrapper = mount(<ProgressControl player={{}} />);
+    expect(wrapper.instance().seekBar).toBeTruthy();
   });
 });