+++ src/index.js
@@ -20,6 +20,7 @@ import MouseTimeDisplay from './components/control-bar/MouseTimeDisplay';
 import VolumeMenuButton from './components/control-bar/VolumeMenuButton';
 import PlaybackRateMenuButton from './components/control-bar/PlaybackRateMenuButton';
 import PlaybackRate from './components/control-bar/PlaybackRate';
+import ClosedCaptionButton from './components/control-bar/ClosedCaptionButton';
 
 import RemainingTimeDisplay from './components/time-controls/RemainingTimeDisplay';
 import CurrentTimeDisplay from './components/time-controls/CurrentTimeDisplay';
@@ -57,6 +58,7 @@ export {
   TimeDivider,
   VolumeMenuButton,
   PlaybackRateMenuButton,
+  ClosedCaptionButton,
   PlaybackRate,
   MenuButton,
   playerActions,